requirements.txt
# Core data
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0

# Machine learning
xgboost>=2.0.0
scikit-learn>=1.3.0
shap>=0.44.0

# Graph
neo4j>=5.0.0
python-dotenv>=1.0.0

# Visualization
matplotlib>=3.7.0
streamlit>=1.30.0
pyvis>=0.3.2

# Experiment tracking (planned)
mlflow>=2.10.0
